One file is enough.
